redact ftdc data

+++ build.sh
@@ -17,12 +17,14 @@ fi
 $DEP ensure $UPDATE
 
 export version="$(git symbolic-ref --short HEAD)-$(date "+%Y%m%d.%s")"
-export ver="1.2.0"
+export ver="1.2.1"
 export version="v${ver}-$(date "+%Y%m%d")"
 mkdir -p build
 env GOOS=linux GOARCH=amd64 go build -ldflags "-X main.version=$version" -o build/keyhole-linux-x64 keyhole.go
 env GOOS=darwin GOARCH=amd64 go build -ldflags "-X main.version=$version" -o build/keyhole-osx-x64 keyhole.go
 env GOOS=windows GOARCH=amd64 go build -ldflags "-X main.version=$version" -o build/keyhole-win-x64.exe keyhole.go
 
-docker build -t simagix/keyhole:latest -t simagix/keyhole:${ver} .
-docker rmi -f $(docker images -f "dangling=true" -q)
+if [ "$1" == "docker" ]; then
+    docker build -t simagix/keyhole:latest -t simagix/keyhole:${ver} .
+    docker rmi -f $(docker images -f "dangling=true" -q)
+fi
